+++ vlm/backends/cpu/src/vlm_backend_cpu.cpp
@@ -12,6 +12,7 @@
 #include <algorithm> // std::fill
 #include <cstdio> // std::printf
 
+#include <stdint.h>
 #include <taskflow/algorithm/for_each.hpp>
 
 #include <lapacke.h>
@@ -22,7 +23,7 @@ using namespace vlm;
 BackendCPU::~BackendCPU() = default; // Destructor definition
 
 BackendCPU::BackendCPU(Mesh& mesh) : Backend(mesh) {
-    lhs.resize((u64)mesh.nb_panels_wing() * (u64)mesh.nb_panels_wing());
+    lhs.resize(mesh.nb_panels_wing() * mesh.nb_panels_wing());
     rhs.resize(mesh.nb_panels_wing());
     ipiv.resize(mesh.nb_panels_wing());
     gamma.resize(mesh.nb_panels_wing());
@@ -40,8 +41,8 @@ void BackendCPU::compute_delta_gamma() {
     std::copy(gamma.begin(), gamma.begin()+mesh.ns, delta_gamma.begin());
 
     // note: this is efficient as the memory is contiguous
-    for (u32 i = 1; i < mesh.nc; i++) {
-        for (u32 j = 0; j < mesh.ns; j++) {
+    for (u64 i = 1; i < mesh.nc; i++) {
+        for (u64 j = 0; j < mesh.ns; j++) {
             delta_gamma[i*mesh.ns + j] = gamma[i*mesh.ns + j] - gamma[(i-1)*mesh.ns + j];
         }
     }
@@ -58,23 +59,23 @@ void BackendCPU::compute_lhs(const FlowData& flow) {
         {m.normal.x.data(), m.normal.y.data(), m.normal.z.data()}
     };
 
-    const u32 start_wing = 0;
-    const u32 end_wing = (m.nc - 1) * m.ns;
+    const u64 start_wing = 0;
+    const u64 end_wing = (m.nc - 1) * m.ns;
 
     tf::Taskflow taskflow;
 
     auto init = taskflow.placeholder();
-    auto wing_pass = taskflow.for_each_index(start_wing, end_wing, [&] (u32 i) {
+    auto wing_pass = taskflow.for_each_index(start_wing, end_wing, [&] (u64 i) {
         ispc::kernel_influence(mesh_proxy, lhs.data(), i, i, flow.sigma_vatistas);
     });
 
-    u32 idx = m.nc - 1;
+    u64 idx = m.nc - 1;
     auto cond = taskflow.emplace([&]{
         return idx < m.nc + m.nw ? 0 : 1; // 0 means continue, 1 means break
     });
-    auto wake_pass = taskflow.for_each_index(0u, m.ns, [&] (u32 j) {
-        const u32 ia = (m.nc - 1) * m.ns + j;
-        const u32 lidx = idx * m.ns + j;
+    auto wake_pass = taskflow.for_each_index(0ull, m.ns, [&] (u64 j) {
+        const u64 ia = (m.nc - 1) * m.ns + j;
+        const u64 lidx = idx * m.ns + j;
         ispc::kernel_influence(mesh_proxy, lhs.data(), ia, lidx, flow.sigma_vatistas);
     });
     auto back = taskflow.emplace([&]{
@@ -92,8 +93,8 @@ void BackendCPU::compute_lhs(const FlowData& flow) {
     Executor::get().run(taskflow).wait();
 }
 
-void kernel_cpu_rhs(u32 n, const float normal_x[], const float normal_y[], const float normal_z[], float freestream_x, float freestream_y, float freestream_z, float rhs[]) {
-    for (u32 i = 0; i < n; i++) {
+void kernel_cpu_rhs(u64 n, const float normal_x[], const float normal_y[], const float normal_z[], float freestream_x, float freestream_y, float freestream_z, float rhs[]) {
+    for (u64 i = 0; i < n; i++) {
         rhs[i] = - (freestream_x * normal_x[i] + freestream_y * normal_y[i] + freestream_z * normal_z[i]);
     }
 }
@@ -109,9 +110,9 @@ void BackendCPU::compute_rhs(const FlowData& flow, const std::vector<f32>& secti
     tiny::ScopedTimer timer("Rebuild RHS");
     assert(section_alphas.size() == mesh.ns);
     const Mesh& m = mesh;
-    for (u32 i = 0; i < mesh.nc; i++) {
-        for (u32 j = 0; j < mesh.ns; j++) {
-            const u32 li = i * mesh.ns + j; // linear index
+    for (u64 i = 0; i < mesh.nc; i++) {
+        for (u64 j = 0; j < mesh.ns; j++) {
+            const u64 li = i * mesh.ns + j; // linear index
             const linalg::alias::float3 freestream = compute_freestream(flow.u_inf, section_alphas[j], flow.beta);
             rhs[li] = - (freestream.x * m.normal.x[li] + freestream.y * m.normal.y[li] + freestream.z * m.normal.z[li]);
         }
@@ -120,28 +121,28 @@ void BackendCPU::compute_rhs(const FlowData& flow, const std::vector<f32>& secti
 
 void BackendCPU::lu_factor() {
     tiny::ScopedTimer timer("Factor");
-    const u32 n = mesh.nb_panels_wing();
+    const int32_t n = static_cast<int32_t>(mesh.nb_panels_wing());
     LAPACKE_sgetrf(LAPACK_COL_MAJOR, n, n, lhs.data(), n, ipiv.data());
 }
 
 void BackendCPU::lu_solve() {
     tiny::ScopedTimer timer("Solve");
-    const u32 n = mesh.nb_panels_wing();
+    const int32_t n = static_cast<int32_t>(mesh.nb_panels_wing());
     std::copy(rhs.begin(), rhs.end(), gamma.begin());
 
     LAPACKE_sgetrs(LAPACK_COL_MAJOR, 'N', n, 1, lhs.data(), n, ipiv.data(), gamma.data(), n);
 }
 
 // f32 BackendCPU::compute_coefficient_cl(const FlowData& flow, const f32 area,
-//     const u32 j, const u32 n) {
+//     const u64 j, const u64 n) {
 //     assert(n > 0);
 //     assert(j >= 0 && j+n <= mesh.ns);
     
 //     f32 cl = 0.0f;
 
-//     for (u32 u = 0; u < mesh.nc; u++) {
-//         for (u32 v = j; v < j + n; v++) {
-//             const u32 li = u * mesh.ns + v; // linear index
+//     for (u64 u = 0; u < mesh.nc; u++) {
+//         for (u64 v = j; v < j + n; v++) {
+//             const u64 li = u * mesh.ns + v; // linear index
 //             const linalg::alias::float3 v0 = mesh.get_v0(li);
 //             const linalg::alias::float3 v1 = mesh.get_v1(li);
 //             // Leading edge vector pointing outward from wing root
@@ -160,16 +161,16 @@ linalg::alias::float3 BackendCPU::compute_coefficient_cm(
     const FlowData& flow,
     const f32 area,
     const f32 chord,
-    const u32 j,
-    const u32 n)
+    const u64 j,
+    const u64 n)
 {
     assert(n > 0);
     assert(j >= 0 && j+n <= mesh.ns);
     linalg::alias::float3 cm(0.f, 0.f, 0.f);
 
-    for (u32 u = 0; u < mesh.nc; u++) {
-        for (u32 v = j; v < j + n; v++) {
-            const u32 li = u * mesh.ns + v; // linear index
+    for (u64 u = 0; u < mesh.nc; u++) {
+        for (u64 v = j; v < j + n; v++) {
+            const u64 li = u * mesh.ns + v; // linear index
             const linalg::alias::float3 v0 = mesh.get_v0(li);
             const linalg::alias::float3 v1 = mesh.get_v1(li);
             // Leading edge vector pointing outward from wing root
@@ -188,8 +189,8 @@ linalg::alias::float3 BackendCPU::compute_coefficient_cm(
 f32 BackendCPU::compute_coefficient_cd(
     const FlowData& flow,
     const f32 area,
-    const u32 j,
-    const u32 n) 
+    const u64 j,
+    const u64 n) 
 {
     tiny::ScopedTimer timer("Compute CD");
     assert(n > 0);
@@ -212,8 +213,8 @@ f32 BackendCPU::compute_coefficient_cd(
 f32 BackendCPU::compute_coefficient_cl(
     const FlowData& flow,
     const f32 area,
-    const u32 j,
-    const u32 n) 
+    const u64 j,
+    const u64 n) 
 {
     Mesh& m = mesh;
     ispc::MeshProxy mesh_proxy = {
